requirements.txt
requests==2.31.0
apache-airflow==3.0.6
apache-airflow-providers-postgres==5.10.0
psycopg2-binary==2.9.9
pendulum==3.0.0